# crates/catalog-agent/src/http/peer_catalog/mod.rs
/*
 * Copyright (C) 2026 - Universidad Politécnica de Madrid - UPM
 *
 * This program is free software: you can redistribute it and/or modify
 * it under the terms of the GNU General Public License as published by
 * the Free Software Foundation, either version 3 of the License, or
 * (at your option) any later version.
 *
 * This program is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
 * GNU General Public License for more details.
 *
 * You should have received a copy of the GNU General Public License
 * along with this program. If not, see <https://www.gnu.org/licenses/>.
 */

use crate::entities::catalogs::{EditCatalogDto, NewCatalogDto};
use crate::http::common::to_camel_case::ToCamelCase;
use crate::services::catalogs::CatalogServiceTrait;
use crate::services::peer_catalogs::PeerCatalogServiceTrait;
use axum::extract::rejection::JsonRejection;
use axum::extract::{FromRef, Path, Query, State};
use axum::response::IntoResponse;
use axum::routing::{delete, get, post, put};
use axum::{Json, Router};
use common::auth::AccessScope;
use common::batch_requests::BatchRequests;
use common::config::services::CatalogConfig;
use common::errors::CommonErrors;
use reqwest::StatusCode;
use serde::Deserialize;
use std::sync::Arc;
use ymir::errors::Outcome;

#[derive(Clone)]
pub struct PeerCatalogEntityRouter {
    service: Arc<dyn PeerCatalogServiceTrait>,
}

impl FromRef<PeerCatalogEntityRouter> for Arc<dyn PeerCatalogServiceTrait> {
    fn from_ref(state: &PeerCatalogEntityRouter) -> Self {
        state.service.clone()
    }
}

impl PeerCatalogEntityRouter {
    pub fn new(service: Arc<dyn PeerCatalogServiceTrait>) -> Self {
        Self { service }
    }

    pub fn router(self) -> Router {
        Router::new()
            .route("/", get(Self::handle_get_all_catalog_by_peer_id))
            .route("/{peer_id}", get(Self::handle_get_catalog_by_peer_id))
            .with_state(self)
    }

    async fn handle_get_all_catalog_by_peer_id(
        State(state): State<PeerCatalogEntityRouter>,
        scope: AccessScope,
    ) -> impl IntoResponse {
        match state.service.get_all_peer_catalogs(&scope).await {
            Ok(data) => (StatusCode::OK, Json(data)).into_response(),
            Err(e) => return e.into_response(),
        }
    }

    async fn handle_get_catalog_by_peer_id(
        State(state): State<PeerCatalogEntityRouter>,
        scope: AccessScope,
        Path(peer_id): Path<String>,
    ) -> impl IntoResponse {
        match state.service.get_peer_catalog(&scope, &peer_id).await {
            Ok(Some(catalog)) => (StatusCode::OK, Json(catalog)).into_response(),
            Ok(None) => {
                let err =
                    CommonErrors::missing_resource_new("peer catalog", "Peer Catalog not found");
                err.into_response()
            }
            Err(e) => return e.into_response(),
        }
    }
}