# static/specs/openapi/fe-gateway.yaml
openapi: 3.1.0
info:
  title: Eunomia DS Frontend Gateway API
  description: "OpenAPI specification for the Eunomia DS-Agent Frontend Gateway.

    The gateway proxies requests from the GUI to backend microservices.

    All paths are prefixed with `/api` and routed by service prefix to the

    corresponding microservice (catalog-agent, negotiation-agent, transfer-agent,
    auth-agent).

    "
  version: 1.0.0
servers:
  - url: "{gateway_url}/api"
    description: Frontend Gateway
    variables:
      gateway_url:
        default: http://localhost:8080
...
paths:
  /fe-config:
    get:
      operationId: getFeConfig
      summary: Get frontend configuration
      tags:
        - Gateway
      responses:
        "200":
          description: Gateway host/port configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeConfig"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorInfo"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorInfo"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorInfo"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorInfo"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorInfo"
...
  /catalogs:
    get:
      operationId: getCatalogs
      summary: List catalogs
      tags:
        - Catalogs
      parameters:
        - name: with_main_catalog
          in: query
          schema:
            type: boolean
            default: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CatalogDto"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorInfo"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorInfo"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorInfo"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorInfo"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorInfo"
    post:
      operationId: createCatalog
      summary: Create a catalog
      tags:
        - Catalogs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewCatalogDto"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CatalogDto"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorInfo"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorInfo"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorInfo"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorInfo"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorInfo"
...
  /catalogs/{catalogId}:
    get:
      operationId: getCatalogById
      summary: Get catalog by ID
      tags:
        - Catalogs
      parameters:
        - $ref: "#/components/parameters/catalogId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CatalogDto"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorInfo"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorInfo"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorInfo"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorInfo"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorInfo"
    put:
      operationId: updateCatalog
      summary: Update a catalog
      tags:
        - Catalogs
      parameters:
        - $ref: "#/components/parameters/catalogId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CatalogDto"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CatalogDto"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorInfo"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorInfo"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorInfo"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorInfo"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorInfo"
    delete:
      operationId: deleteCatalog
      summary: Delete a catalog
      tags:
        - Catalogs
      parameters:
        - $ref: "#/components/parameters/catalogId"
      responses:
        "204":
          description: Deleted
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorInfo"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorInfo"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorInfo"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorInfo"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorInfo"
...
  /peer-catalogs:
    get:
      tags:
        - Peer Catalogs
      summary: Get all remote peer catalog
      operationId: getAllPeerCatalog
      responses:
        "200":
          description: Peer Catalogs found
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PeerCatalogEntry"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorInfo"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorInfo"
  /peer-catalogs/{peerId}:
    get:
      operationId: getPeerCatalog
      summary: Get remote peer catalog
      tags:
        - Catalogs
      parameters:
        - name: peerId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Catalog"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorInfo"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorInfo"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorInfo"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorInfo"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorInfo"
...
  /datasets/catalog/{catalogId}:
    get:
      operationId: getDatasetsByCatalogId
      summary: List datasets by catalog
      tags:
        - Datasets
      parameters:
        - $ref: "#/components/parameters/catalogId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DatasetDto"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorInfo"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorInfo"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorInfo"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorInfo"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorInfo"
...
  /data-services/catalog/{catalogId}:
    get:
      operationId: getDataServicesByCatalogId
      summary: List data services by catalog
      tags:
        - DataServices
      parameters:
        - $ref: "#/components/parameters/catalogId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DataServiceDto"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorInfo"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorInfo"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorInfo"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorInfo"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorInfo"
...
  /transfers/dataplane/dataplane-processes/{dataplaneProcessId}/logs:
    get:
      tags:
        - Dataplane Logs
      summary: Get dataplane transfer logs
      operationId: getDataplaneTransferLogs
      parameters:
        - name: dataplaneProcessId
          in: path
          required: true
          schema:
            type: string
            format: urn
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DataplaneTransferLogDto"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /did-json/{url}:
    get:
      operationId: getDidFromUrl
      summary: Get DID JSON from URL
      tags:
        - DID
      parameters:
        - $ref: "#/components/parameters/url"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Did"
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
...
components:
  parameters:
...
    catalogId:
      name: catalogId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    url:
      name: url
      in: path
      required: true
      schema:
        type: string
...
  schemas:
...
    FeConfig:
      type: object
      properties:
        gateway_host:
          type: string
        gateway_port:
          type: string
    ErrorInfo:
      type: object
      required:
        - message
        - error_code
        - cause
      properties:
        message:
          type: string
        error_code:
          type: integer
        details:
          type: string
        cause:
          type: string
      examples:
        - message: Resource not found
          error_code: 3200
          status_code: 404
          cause: Catalog with ID urn:uuid:123 not found
...
    CatalogDto:
      type: object
      properties:
        id:
          type: string
          format: urn
        foafHomePage:
          type: string
        dctConformsTo:
          type: string
        dctCreator:
          type: string
        dctIdentifier:
          type: string
        dctIssued:
          type: string
          format: date-time
        dctModified:
          type: string
          format: date-time
        dctTitle:
          type: string
        dspaceParticipantId:
          type: string
        dspaceMainCatalog:
          type: boolean
      examples:
        - dctConformsTo: null
          dctCreator: null
          dctIdentifier: urn:catalog:35d10981-1d0b-4204-b32e-847b827ce635
          dctIssued: "2026-01-09T10:49:53.822684Z"
          dctModified: null
          dctTitle: null
          dspaceMainCatalog: true
          dspaceParticipantId: did:jwk:eyJrdHkiOiJSU0Ei...
          foafHomePage: null
          id: urn:catalog:35d10981-1d0b-4204-b32e-847b827ce635
    NewCatalogDto:
      type: object
      properties:
        id:
          type: string
          format: urn
        foafHomePage:
          type: string
        dctConformsTo:
          type: string
        dctCreator:
          type: string
        dctTitle:
          type: string
        dspaceParticipantId:
          type: string
      examples:
        - dctConformsTo: string
          dctCreator: string
          dctTitle: string
          foafHomePage: string
    DatasetDto:
      type: object
      properties:
        id:
          type: string
          format: urn
        dctConformsTo:
          type: string
        dctCreator:
          type: string
        dctIdentifier:
          type: string
        dctIssued:
          type: string
          format: date-time
        dctModified:
          type: string
          format: date-time
        dctTitle:
          type: string
        dctDescription:
          type: string
        catalogId:
          type: string
          format: urn
      examples:
        - catalogId: urn:catalog:4958f37d-17cf-44e7-ae18-1f26b6fe15ff
          dctConformsTo: null
          dctCreator: null
          dctDescription: null
          dctIdentifier: urn:dataset:10825ca3-a75e-43bd-9a82-603759cb8dec
          dctIssued: "2026-01-10T14:23:29.091150Z"
          dctModified: null
          dctTitle: superfoaf
          id: urn:dataset:10825ca3-a75e-43bd-9a82-603759cb8dec
...
    DataServiceDto:
      type: object
      properties:
        id:
          type: string
          format: urn
        dcatEndpointDescription:
          type: string
        dcatEndpointUrl:
          type: string
        dctConformsTo:
          type: string
        dctCreator:
          type: string
        dctIdentifier:
          type: string
        dctIssued:
          type: string
          format: date-time
        dctModified:
          type: string
          format: date-time
        dctTitle:
          type: string
        dctDescription:
          type: string
        catalogId:
          type: string
          format: urn
        dspaceMainDataService:
          type: boolean
      examples:
        - catalogId: urn:catalog:35d10981-1d0b-4204-b32e-847b827ce635
          dcatEndpointDescription: null
          dcatEndpointUrl: http://127.0.0.1:1200/dsp/current
          dctConformsTo: null
          dctCreator: null
          dctDescription: null
          dctIdentifier: urn:data-service:37840ca3-79f6-4542-8394-0c57dd93650f
          dctIssued: "2026-01-09T10:49:53.895893Z"
          dctModified: null
          dctTitle: null
          dspaceMainDataService: true
          id: urn:data-service:37840ca3-79f6-4542-8394-0c57dd93650f
...
    OdrlPermission:
      type: object
      required:
        - action
      properties:
        action:
          type: string
        constraint:
          type: array
          items:
            $ref: "#/components/schemas/OdrlConstraint"
        duty:
          type: array
          items:
            $ref: "#/components/schemas/OdrlDuty"
...
    Catalog:
      type: object
      properties:
        "@context":
          oneOf:
            - type: string
            - type: array
              items:
                type: string
        "@type":
          type: string
        "@id":
          type: string
          format: urn
        homepage:
          type: string
        theme:
          type: string
        keyword:
          type: string
        conformsTo:
          type: string
        creator:
          type: string
        identifier:
          type: string
        issued:
          type: string
          format: date-time
        modified:
          type: string
          format: date-time
        title:
          type: string
        description:
          type: array
          items:
            type: string
        participantId:
          type: string
        hasPolicy:
          type: array
          items:
            $ref: "#/components/schemas/OdrlOffer"
        extraFields:
          type: object
        catalog:
          oneOf:
            - type: array
              items:
                $ref: "#/components/schemas/Catalog"
            - type: array
              items:
                $ref: "#/components/schemas/CatalogMinimized"
        dataset:
          oneOf:
            - type: array
              items:
                $ref: "#/components/schemas/Dataset"
            - type: array
              items:
                $ref: "#/components/schemas/DatasetMinimized"
        service:
          oneOf:
            - $ref: "#/components/schemas/DataService"
            - $ref: "#/components/schemas/DataServiceMinimized"
            - type: array
              items:
                $ref: "#/components/schemas/DataService"
      examples:
        - "@context":
            - https://w3id.org/dspace/2025/1/context.jsonld
          "@type": Catalog
          "@id": urn:catalog:35d10981-1d0b-4204-b32e-847b827ce635
          identifier: urn:catalog:35d10981-1d0b-4204-b32e-847b827ce635
          issued: "2026-01-09T10:49:53.822684"
          catalog:
            - "@type": Catalog
              "@id": urn:catalog:4958f37d-17cf-44e7-ae18-1f26b6fe15ff
              homepage: superfoaf
              identifier: urn:catalog:4958f37d-17cf-44e7-ae18-1f26b6fe15ff
              issued: "2026-01-10T14:23:16.619801"
              dataset:
                - "@type": Dataset
                  "@id": urn:dataset:10825ca3-a75e-43bd-9a82-603759cb8dec
                  identifier: urn:dataset:10825ca3-a75e-43bd-9a82-603759cb8dec
                  issued: "2026-01-10T14:23:29.091150"
                  title: superfoaf
                  hasPolicy:
                    - "@id": urn:odrl-policy:3953b047-0aa5-4649-a0bd-0522f77d018b
                      permission:
                        - action: use
                      "@type": Offer
                  distribution:
                    - "@type": Distribution
                      "@id": urn:distribution:77116828-c4dd-4632-8946-969bb7903819
                      accessService:
                        "@type": DataService
                        "@id": urn:data-service:37840ca3-79f6-4542-8394-0c57dd93650f
                        endpointURL: http://127.0.0.1:1200/dsp/current
                      issued: "2026-01-10T14:24:29.424022"
                      title: superfoaf
                      formats: http+push
              service: []
          dataset: []
          service:
            "@type": DataService
            "@id": urn:data-service:37840ca3-79f6-4542-8394-0c57dd93650f
            endpointURL: http://127.0.0.1:1200/dsp/current
    Dataset:
      type: object
      properties:
        "@context":
          oneOf:
            - type: string
            - type: array
              items:
                type: string
        "@type":
          type: string
        "@id":
          type: string
        hasPolicy:
          type: array
          items:
            $ref: "#/components/schemas/OdrlOffer"
        distribution:
          oneOf:
            - type: array
              items:
                $ref: "#/components/schemas/Distribution"
            - type: array
              items:
                $ref: "#/components/schemas/DistributionMinimized"
        conformsTo:
          type: string
        title:
          type: string
      examples:
        - "@context":
            - https://w3id.org/dspace/2025/1/context.jsonld
          "@type": Dataset
          "@id": urn:dataset:10825ca3-a75e-43bd-9a82-603759cb8dec
          identifier: urn:dataset:10825ca3-a75e-43bd-9a82-603759cb8dec
          issued: "2026-01-10T14:23:29.091150"
          title: superfoaf
          hasPolicy:
            - "@id": urn:odrl-policy:3953b047-0aa5-4649-a0bd-0522f77d018b
              permission:
                - action: use
              "@type": Offer
          distribution:
            - "@type": Distribution
              "@id": urn:distribution:77116828-c4dd-4632-8946-969bb7903819
              accessService:
                "@type": DataService
                "@id": urn:data-service:37840ca3-79f6-4542-8394-0c57dd93650f
                endpointURL: http://127.0.0.1:1200/dsp/current
              issued: "2026-01-10T14:24:29.424022"
              title: superfoaf
              formats: http+push
...
    DataService:
      type: object
      properties:
        "@context":
          oneOf:
            - type: string
            - type: array
              items:
                type: string
        "@type":
          type: string
        "@id":
          type: string
        endpointURL:
          type: string
        endpointDescription:
          type: string
    OdrlOffer:
      type: object
      properties:
        "@type":
          type: string
          enum:
            - Offer
            - Agreement
        "@id":
          type: string
          format: urn
        target:
          type: string
          format: urn
        permission:
          type: array
          items:
            $ref: "#/components/schemas/OdrlPermission"
        prohibition:
          type: array
          items:
            $ref: "#/components/schemas/OdrlProhibition"
        obligation:
          type: array
          items:
            $ref: "#/components/schemas/OdrlObligation"
...
    DataplaneTransferLogDto:
      type: object
      properties:
        id:
          type: string
          format: urn
        dataplane_process_id:
          type: string
          format: urn
        previous_state:
          type:
            - string
            - "null"
        new_state:
          type: string
        trigger:
          type: string
        reason:
          type:
            - string
            - "null"
        created_at:
          type: string
          format: date-time
...
    ParticipantDto:
      type: object
      properties:
        participant_id:
          type: string
        participant_nick:
          type: string
        participant_type:
          type: string
        base_url:
          type: string
        token:
          type: string
        saved_at:
          type: string
          format: date-time
        last_interaction:
          type: string
          format: date-time
        extra_fields:
          type: object
          additionalProperties: true
    PeerCatalogEntry:
      type: object
      properties:
        participant:
          $ref: "#/components/schemas/ParticipantDto"
        catalog:
          $ref: "#/components/schemas/Catalog"
...
    CatalogMinimized:
      type: object
      properties:
        "@type":
          type: string
        "@id":
          type: string
          format: urn
        homepage:
          type: string
    DatasetMinimized:
      type: object
      properties:
        "@type":
          type: string
        "@id":
          type: string
    DataServiceMinimized:
      type: object
      properties:
        "@type":
          type: string
        "@id":
          type: string
...
    Did:
      type: object
      required:
        - "@context"
        - assertionMethod
        - authentication
        - capabilityDelegation
        - capabilityInvocation
        - id
        - keyAgreement
        - verificationMethod
      title: Did
      properties:
        "@context":
          type: array
          items:
            type: string
        assertionMethod:
          type: array
          items:
            type: string
        authentication:
          type: array
          items:
            type: string
        capabilityDelegation:
          type: array
          items:
            type: string
        capabilityInvocation:
          type: array
          items:
            type: string
        id:
          type: string
        keyAgreement:
          type: array
          items:
            type: string
        verificationMethod:
          type: array
          items:
            "$ref": "#/components/schemas/DidVerificationMethod"
    DidVerificationMethod:
      type: object
      title: VerificationMethod
      required:
        - controller
        - id
        - publicKeyJwk
        - type
      properties:
        controller:
          type: string
        id:
          type: string
        publicKeyJwk:
          type: object
        type:
          type: string
...
  responses:
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorInfo"
...